requirements.txt
scipion-pyworkflow
codespeed-client
